+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -20,6 +20,7 @@ import org.mockito.Mockito.mock
 import org.mockito.Mockito.verify
 import org.mozilla.geckoview.GeckoRuntime
 import org.mozilla.geckoview.GeckoRuntimeSettings
+import org.mozilla.geckoview.GeckoWebExecutor
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 
@@ -110,5 +111,16 @@ class GeckoEngineTest {
         verify(runtimeSettings).trackingProtectionCategories = TrackingProtectionPolicy.all().categories
         assertTrue(engine.settings.testingModeEnabled)
         assertEquals("test-ua", engine.settings.userAgentString)
+    }
+
+    @Test
+    fun `speculativeConnect forwards call to executor`() {
+        val executor: GeckoWebExecutor = mozilla.components.support.test.mock()
+
+        val engine = GeckoEngine(context, runtime = runtime, executorProvider = { executor })
+
+        engine.speculativeConnect("https://www.mozilla.org")
+
+        verify(executor).speculativeConnect("https://www.mozilla.org")
     }
 }