+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -8,6 +8,7 @@ import android.content.Context
 import mozilla.components.concept.engine.DefaultSettings
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy
 import mozilla.components.concept.engine.UnsupportedSettingException
+import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
@@ -20,6 +21,7 @@ import org.mockito.Mockito.mock
 import org.mockito.Mockito.verify
 import org.mozilla.geckoview.GeckoRuntime
 import org.mozilla.geckoview.GeckoRuntimeSettings
+import org.mozilla.geckoview.GeckoWebExecutor
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 
@@ -110,5 +112,16 @@ class GeckoEngineTest {
         verify(runtimeSettings).trackingProtectionCategories = TrackingProtectionPolicy.all().categories
         assertTrue(engine.settings.testingModeEnabled)
         assertEquals("test-ua", engine.settings.userAgentString)
+    }
+
+    @Test
+    fun `speculativeConnect forwards call to executor`() {
+        val executor: GeckoWebExecutor = mock()
+
+        val engine = GeckoEngine(context, runtime = runtime, executorProvider = { executor })
+
+        engine.speculativeConnect("https://www.mozilla.org")
+
+        verify(executor).speculativeConnect("https://www.mozilla.org")
     }
 }