+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -45,6 +45,13 @@ class GeckoEngine(
         return GeckoEngineSessionState.fromJSON(json)
     }
 
+    /**
+     * Opens a speculative connection to the host of [url].
+     *
+     * Note: This implementation is a no-op.
+     */
+    override fun speculativeConnect(url: String) = Unit
+
     override fun name(): String = "Gecko"
 
     /**