+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngine.kt
@@ -48,6 +48,13 @@ class SystemEngine(
     }
 
     /**
+     * Opens a speculative connection to the host of [url].
+     *
+     * Note: This implementation is a no-op.
+     */
+    override fun speculativeConnect(url: String) = Unit
+
+    /**
      * See [Engine.name]
      */
     override fun name(): String = "System"