+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Engine.kt
@@ -47,6 +47,16 @@ interface Engine {
     fun name(): String
 
     /**
+     * Opens a speculative connection to the host of [url].
+     *
+     * This is useful if an app thinks it may be making a request to that host in the near future. If no request
+     * is made, the connection will be cleaned up after an unspecified.
+     *
+     * Not all [Engine] implementations may actually implement this.
+     */
+    fun speculativeConnect(url: String)
+
+    /**
      * Provides access to the settings of this engine.
      */
     val settings: Settings